Bug fix related to size_t
Former-commit-id: a2e704821640830752544bc89b2c3efea415a162

+++ applications/programs/mpi_classify_CL2D/mpi_classify_CL2D_main.cpp
@@ -1032,7 +1032,7 @@ void CL2D::run(const FileName &fnODir, const FileName &fnOut, int level)
 
             std::cout << "Number of assignment changes=" << Nchanges
             << std::endl;
-            MDChanges.setValue(MDL_CL2D_CHANGES, Nchanges, idMdChanges);
+            MDChanges.setValue(MDL_CL2D_CHANGES, (int)Nchanges, idMdChanges);
             MDChanges.write(formatString("info@%s",fnClasses.c_str()),MD_APPEND);
         }
